Update expanding_biology_database.json
Added a few more papers. 
Added hash property. The python script now checks whether the pdf has already been parsed first using a calculated hash (that particular identical PDF) and then based on DOI (for example in case a preprint --> published paper). If the hash is different but the DOI exists, the entry will be overwritten and replaced with the new version.